requirements.txt
# Core dependencies
llama-cpp-python>=0.2.50
pyyaml>=6.0
requests>=2.28.0
pydantic>=2.0
ddgs>=9.0.0
aiohttp>=3.9.0
datasets>=2.19.0
